package/requirements.txt
# endpoint reqs
attrs
pyyaml
requests
python-dateutil
redis

# diana reqs
pydicom
numpy
click
